Move login command to internal
