+++ internal/pkg/cmd/common/util.go
